relocate get started accordion to just the component it belongs to
